Moved ContextEntry class from surfactant/sbomtypes/ into surfactant/config.py
